[TODO] implement get-pstat
