+++ user/getpinfo.c
@@ -1,4 +1,10 @@
 #include "kernel/types.h"
 #include "kernel/stat.h"
+#include "kernel/pstat.h"
 #include "user/user.h"
 
+void
+main(void) {
+  extern struct pinfo;
+
+}
